Basic Draft of DocReg is up. DBConfig now prints SQL errors to our home, so we know of any SQL errors that happen.
InpModder allows us to clean up input and remove trailing slashes and stuff. Some Refactoring in our index.php page that missed me. DocReg checks for any duplicates (in both tables) before making an insert. It saves a hashed value for the password into the database. I am using a basic on at the moment and will move to better ones if possible.