Prescribe Medication is fully up. As of now, the toughest of the back is finished.
DeletePrescData.js is the function bind for the delete button on our page.

ValidationStuff.js check-up on data that is entered and makes sure it is filled in correctly.

AddElems requires no CSS. It accepts post data and adds that medicine to out table whereas DelElems deletes an existing medicine.

DispPrescDoc.php is the function the retrieves the prescription entered so far and displays it. It hold JQuery for the delete part as it requires another sectional reload otherwise.

FillPrescription.php is where all the doctor enters and sees all of the stuff.

The other two have minor changes and save session variables to allow communication across pages.