The Set-Up to View a prescription and associated data is finished.
Login now saves a session variable and performs a redirect

PrescList shows us our Prescriptions for a user.

PrescPage.php is where we'll show the medication. This is unfinished. Expect it by tomorrow the latest.

ViewPrescription.js holds the function bind for our buttons.

SetUpPresc saves our Prescription ID as a session Variable instead of a post value. This is easier because the other option is to make a form and figure out which button was clicked and then pick that value. We could even just go back and display medication on the same page if we want to. This method gives us options.