ViewHistory.php is up
So, in the previous commit I added an extra value to our JSON post request to render a different page for different data. This commit has our FindPat page to Find a patient so we can look up their history. I need to finish up ViewHistory page. Have it link to view prescription when they click on " *View More Details* " or something like that.