View History is fully up
I set up a separate JS file that works similar to when a patient wants to view logs. I tried it in the same file but for some reason, it wouldn't work so I made a separate one instead and it just works somehow. I have tested and formatted the back buttons in the ViewDetailedHistory page and set-up everything accordingly. This adds one more file to set up the `css` for apart from ViewHistory.

The gist of what happens is as follows:
>You Find a patient
>You confirm the patient and that value is saved in a session variable
>We show all allotted prescriptions for that patient
>You choose to view more info about a certain prescription
>On Clicking the button, we send ID to the JS file
>We forward that value ahead and save it in our session variable
>We redirect to `ViewDetailedHistory.php` where you can see the entire prescription, even if it's expired.